requirements.txt
Django==2.0.5
django-crispy-forms==1.7.2
djangorestframework==3.8.2
markdown==2.6.11
django-filter==1.1.0
